fix: Entity PK 값과 JPA 타입이 맞지 않는 것 수정 등
Entity PK 값과 JPA 타입이 맞지 않는 것 수정
Repository 작동 여부 확인

+++ server/server/src/main/java/com/mjubus/server/repository/StationLocationRepository.java
@@ -3,5 +3,5 @@ package com.mjubus.server.repository;
 import com.mjubus.server.domain.StationLocation;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface StationLocationRepository extends JpaRepository<StationLocation, Long> {
+public interface StationLocationRepository extends JpaRepository<StationLocation, Integer> {
 }
